feat: split web build into worker & ui packages
